requirements.txt
certifi==2026.2.25
charset-normalizer==3.4.4
idna==3.11
psycopg==3.3.3
psycopg-binary==3.3.3
python-dotenv==1.2.2
redis==7.2.1
requests==2.32.5
spotipy==2.26.0
urllib3==2.6.3
